pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-juice-kor"
version = "0.1.0"
description = "FinancialJuice SignalR client with translation and Slack integration"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "cslee" }
]
dependencies = [
  "requests>=2.31",
  "websocket-client>=1.7",
  "python-dotenv>=1.0",
  "firebase-admin>=6.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
fj-client = "financial_juice_client:main"
fj-translate = "ai_translator:main"
fj-ingest = "fj_client.ingest_cli:main"

[tool.setuptools]
packages = ["fj_client"]

